requirements.txt
numpy
pandas
setuptools
mysql-connector-python
pymysql
python-dotenv
scikit-learn